separate persistent kernel scheduler from registry
