+++ csrc/scheduler/registry.cpp
@@ -16,925 +16,18 @@
 #include <scheduler/debug_utils.h>
 #include <scheduler/matmul_utils.h>
 #include <scheduler/normalization_utils.h>
+#include <scheduler/persistent_kernel_scheduler.h>
 #include <scheduler/pointwise.h>
 #include <scheduler/registry.h>
+#include <scheduler/registry_utils.h>
 #include <scheduler/transpose.h>
 #include <scheduler/utils.h>
 #include <tensor_metadata.h>
 
+#include <ATen/cuda/CUDAContext.h>
 #include <limits>
 
-#include <ATen/cuda/CUDAContext.h>
-
 namespace nvfuser {
-
-namespace {
-// TODO: Deduplicate from compute_at.cpp
-std::deque<std::deque<TensorView*>> tvChains(
-    std::deque<std::deque<Val*>> val_chains) {
-  std::deque<std::deque<TensorView*>> tv_chains(val_chains.size());
-  for (const auto i : c10::irange(val_chains.size())) {
-    auto tv_iterable = ir_utils::filterByType<TensorView>(val_chains[i]);
-    tv_chains[i] =
-        std::deque<TensorView*>(tv_iterable.begin(), tv_iterable.end());
-  }
-  return tv_chains;
-}
-
-bool rejectScheduleFusionInputRequirement(
-    Expr* expr,
-    ScheduleHeuristic schedule_stragety) {
-  if (!expr->input(0)->isFusionInput()) {
-    scheduler_debug_utils::canScheduleRejectReason(
-        schedule_stragety,
-        "First input of ",
-        expr->getOpString(),
-        " must be fusion input.");
-    return true;
-  }
-  return false;
-}
-
-// TODO: remove this requirement entirely
-bool rejectScheduleForMemoryPromotion(
-    Fusion* fusion,
-    ScheduleHeuristic schedule_strategy) {
-  for (auto expr : fusion->exprs()) {
-    if (expr->isOneOf<SelectOp, IndexSelectOp, TorchGatherOp>()) {
-      // For now, only relax the input requirement when it's
-      // take_along_axis. Also since this would require memory
-      // promotion, i.e., persistent global sync in the case of
-      // block-parallel ops, it needs to be explictly enabled.
-      if (expr->isA<TorchGatherOp>() &&
-          expr->as<TorchGatherOp>()->exactSizes() &&
-          isOptionEnabled(EnableOption::MemoryPromotion)) {
-        continue;
-      }
-      if (rejectScheduleFusionInputRequirement(expr, schedule_strategy)) {
-        return true;
-      }
-    }
-
-    // Similarly, ops based resize, such as like slice, pad and cat,
-    // may require memory promotion. Require them to be done with
-    // fusion inputs unless explicitly allowed
-    if (!isOptionEnabled(EnableOption::MemoryPromotion) &&
-        std::any_of(
-            expr->outputs().begin(), expr->outputs().end(), [](Val* output) {
-              return output->isA<TensorView>() &&
-                  ir_utils::hasResizedRfactor(output->as<TensorView>());
-            })) {
-      if (rejectScheduleFusionInputRequirement(expr, schedule_strategy)) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-class SchedulerTopologyChecker {
- public:
-  // Checks if any broadcasts are resolved after a reduction that don't follow
-  // the normalization pattern
-  static bool hasNonNormalizePostReductionBCast(Fusion* fusion) {
-    auto all_vals = fusion->usedMathVals();
-    std::vector<TensorView*> reduction_tvs;
-    for (auto tv : ir_utils::filterByType<TensorView>(all_vals)) {
-      if (tv->hasReduction() &&
-          !(fusion == tv->fusion() && tv->isFusionInput())) {
-        reduction_tvs.push_back(tv);
-      }
-    }
-
-    // All tensor views that are eventually consumed to produce a reduction,
-    // includes reduction tensor views.
-    std::unordered_set<TensorView*> pre_reduction_tvs;
-
-    {
-      auto pre_reduction_vals = DependencyCheck::getAllValsBetween(
-          {fusion->inputs().begin(), fusion->inputs().end()},
-          {reduction_tvs.begin(), reduction_tvs.end()});
-      auto pre_reduction_tv_vector =
-          ir_utils::filterByType<TensorView>(pre_reduction_vals);
-      pre_reduction_tvs = std::unordered_set<TensorView*>(
-          pre_reduction_tv_vector.begin(), pre_reduction_tv_vector.end());
-    }
-
-    // Track which tensor views we've validated so we don't do it again.
-    std::unordered_set<TensorView*> validated_resolved_tvs;
-
-    // Run forward (towards outputs) from reductions on any path that isn't
-    // before another reduction. Look for resolved broadcasts. If a resolved
-    // broadcast is found, start there and propagate backwards. Track the id's
-    // that were resolved and make sure there's a mapping to a TensorView before
-    // a reduction.
-    for (auto red_tv : reduction_tvs) {
-      auto forward_tv_chains =
-          tvChains(DependencyCheck::getAllUseChains(red_tv));
-      // Propagate forward from reduction through all uses of the reduction
-      for (auto forward_tv_dep_chain : forward_tv_chains) {
-        TensorView* forward_running_producer = nullptr;
-        TensorView* forward_running_consumer = forward_tv_dep_chain.front();
-        forward_tv_dep_chain.pop_front();
-        while (!forward_tv_dep_chain.empty()) {
-          forward_running_producer = forward_running_consumer;
-          forward_running_consumer = forward_tv_dep_chain.front();
-          forward_tv_dep_chain.pop_front();
-
-          if (std::none_of(
-                  forward_running_producer->getMaybeRFactorDomain().begin(),
-                  forward_running_producer->getMaybeRFactorDomain().end(),
-                  [](IterDomain* id) { return id->isBroadcast(); })) {
-            // If there's no broadcast axes in producer it doesn't need to be
-            // checked
-            continue;
-          }
-
-          // If consumer is before another reduction it doesn't need to be
-          // checked
-          if (pre_reduction_tvs.count(forward_running_consumer)) {
-            break;
-          }
-
-          // If consumer was already validated it doesn't need to be checked
-          if (validated_resolved_tvs.count(forward_running_consumer)) {
-            continue;
-          }
-
-          auto forward_pairwise_root_map = PairwiseRootDomainMap(
-              forward_running_producer, forward_running_consumer);
-          auto forward_p2c_root_map =
-              forward_pairwise_root_map.mapProducerToConsumer(
-                  forward_running_producer->domain(),
-                  forward_running_consumer->domain());
-
-          // These are the ids we will have to resolve. As we resolve them we'll
-          // remove them from this vector. If this vector ends up empty, then
-          // we've resolved everything we need to. This is a pair so as we
-          // traverse we can map the id through the traversal. The first entry
-          // in the pair will be the original id so we can reset it if it's not
-          // resolved before the next traversal. The second ID will be
-          // propagated as we map the IDs through the backward traversal.
-          std::vector<std::pair<IterDomain*, IterDomain*>> ids_to_resolve;
-
-          // Check if any TensorViews have a resolved broadcast
-          for (auto entry : forward_p2c_root_map) {
-            auto p_id = entry.first;
-            auto c_id = entry.second;
-            if (p_id->isBroadcast() && !c_id->isBroadcast()) {
-              ids_to_resolve.emplace_back(c_id, c_id);
-            }
-          }
-
-          if (ids_to_resolve.empty()) {
-            continue;
-          }
-
-          // Only because of api limitations in getAllDependencyChains
-          auto inputs_of_forward_running_consumer =
-              IterVisitor::getInputsTo({forward_running_consumer});
-          auto tv_inputs_of_forward_running_consumer =
-              ir_utils::filterByType<TensorView>(
-                  inputs_of_forward_running_consumer);
-
-          for (auto input_of_forward_running_consumer :
-               tv_inputs_of_forward_running_consumer) {
-            if (pre_reduction_tvs.find(input_of_forward_running_consumer) ==
-                pre_reduction_tvs.end()) {
-              // If this input isn't an input to a reduction, no point
-              // traversing the dependency chains as we know we can't validate
-              // this broadcast through chains to this input
-              continue;
-            }
-
-            auto backward_tv_chains =
-                tvChains(DependencyCheck::getAllDependencyChains(
-                    input_of_forward_running_consumer,
-                    forward_running_consumer));
-
-            for (auto backward_tv_chain : backward_tv_chains) {
-              if (ids_to_resolve.empty()) {
-                break;
-              }
-
-              for (auto& pair : ids_to_resolve) {
-                pair.second = pair.first;
-              }
-
-              TensorView* backward_running_producer = backward_tv_chain.back();
-              TensorView* backward_running_consumer = nullptr;
-              backward_tv_chain.pop_back();
-
-              NVF_ERROR(backward_running_producer == forward_running_consumer);
-
-              while (!backward_tv_chain.empty()) {
-                backward_running_consumer = backward_running_producer;
-                backward_running_producer = backward_tv_chain.back();
-                backward_tv_chain.pop_back();
-
-                std::vector<IterDomain*> running_resolved_ids;
-
-                auto backward_pairwise_root_map = PairwiseRootDomainMap(
-                    backward_running_producer, backward_running_consumer);
-
-                auto backward_c2p_root_map =
-                    backward_pairwise_root_map.mapConsumerToProducer(
-                        backward_running_consumer->domain(),
-                        backward_running_producer->domain());
-
-                // Mark if producer is a producer of a reduction
-                bool producer_resolves =
-                    pre_reduction_tvs.count(backward_running_producer);
-
-                bool at_leat_one_id_mapped = false;
-                for (size_t entry_i = ids_to_resolve.size(); entry_i > 0;
-                     entry_i--) {
-                  auto orig_id = ids_to_resolve[entry_i - 1].first;
-                  auto running_id = ids_to_resolve[entry_i - 1].second;
-                  if (backward_c2p_root_map.find(running_id) !=
-                      backward_c2p_root_map.end()) {
-                    at_leat_one_id_mapped = true;
-                    if (producer_resolves &&
-                        !backward_c2p_root_map.at(running_id)->isBroadcast()) {
-                      // If mapped, and producer is a producer of a reduction,
-                      // we can resolve this id
-                      ids_to_resolve.erase(
-                          ids_to_resolve.begin() + (int64_t)entry_i - 1);
-                    } else {
-                      ids_to_resolve[entry_i - 1] = std::make_pair(
-                          orig_id, backward_c2p_root_map.at(running_id));
-                    }
-                  }
-                }
-                if (!at_leat_one_id_mapped) {
-                  // If no id's map any more, go to the next chain
-                  break;
-                }
-
-                if (ids_to_resolve.empty()) {
-                  break;
-                }
-              }
-            }
-          } // for(auto input_of_forward_running_consumer :
-            // tv_inputs_of_forward_running_consumer){
-
-          // if all ids were not resolved, then we've found an instance of a
-          // bad broadcast resolution after reduction
-          if (!ids_to_resolve.empty()) {
-            return true;
-          }
-
-        } // while (!forward_tv_dep_chain.empty()) {
-      } // for (auto forward_tv_dep_chain : forward_tv_chains) {
-    } // for (auto red_tv : reduction_tvs)
-    return false;
-  }
-
-  // Checks if any broadcasts are resolved after a reduction, this shouldn't be
-  // accepted in the single reduction or multi-reduction scheduler
-  static bool hasPostReductionBCast(Fusion* fusion) {
-    auto all_vals = fusion->usedMathVals();
-    for (auto tv : ir_utils::filterByType<TensorView>(all_vals)) {
-      // Reductions can have multiple outputs, so do this on all found reduction
-      // tensor views
-      if (tv->hasReduction() && !tv->isFusionInput()) {
-        auto tv_chains = tvChains(DependencyCheck::getAllUseChains(tv));
-        // Propagate forward from reduction through all uses of the reduction
-        for (auto tv_dep_chain : tv_chains) {
-          TensorView* running_producer = nullptr;
-          TensorView* running_consumer = tv_dep_chain.front();
-          tv_dep_chain.pop_front();
-          while (!tv_dep_chain.empty()) {
-            running_producer = running_consumer;
-            running_consumer = tv_dep_chain.front();
-            tv_dep_chain.pop_front();
-
-            auto pairwise_root_map =
-                PairwiseRootDomainMap(running_producer, running_consumer);
-            auto p2c_root_map = pairwise_root_map.mapProducerToConsumer(
-                running_producer->domain(), running_consumer->domain());
-
-            // Check if any TensorViews have a resolved broadcast
-            for (auto entry : p2c_root_map) {
-              auto p_id = entry.first;
-              auto c_id = entry.second;
-              if (p_id->isBroadcast() && !c_id->isBroadcast()) {
-                return true;
-              }
-            }
-          }
-        }
-      }
-    }
-    return false;
-  }
-
-  // Checks if there's any unsupported operations post reduction. If outer
-  // reduction we can fuse some pointwise ops if they don't require
-  // broadcasting (checked in hasPostReductionBCast). For inner reductions we
-  // cannot fuse any binary like operation (includes operations like shift that
-  // we're not fusing right now) involving "new" inputs (not going through a
-  // reduction).
-  static bool supportedPostReductionFusion(
-      Fusion* fusion,
-      std::vector<TensorView*> reduction_tvs) {
-    NVF_ERROR(!reduction_tvs.empty());
-    bool fastest_dim_reduction = true;
-    auto red_root_dom = reduction_tvs[0]->getRootDomain();
-    for (size_t i = red_root_dom.size(); i > 0; i--) {
-      if (red_root_dom[i - 1]->isBroadcast()) {
-        continue;
-      } else if (red_root_dom[i - 1]->isReduction()) {
-        fastest_dim_reduction = true;
-        break;
-      } else {
-        fastest_dim_reduction = false;
-        break;
-      }
-    }
-
-    // When checking post reduction vals, we need to make sure
-    //  we are really checking paths starting from all outputs
-    //  of multi-output reductions, i.e. welford/grouped reduction. The
-    //  reduction_tv vector is assumed to only have one of them.
-    std::unordered_set<Val*> reduction_tv_set(
-        reduction_tvs.begin(), reduction_tvs.end());
-
-    for (auto red : reduction_tvs) {
-      if (red->definition()) {
-        if (ir_utils::isReductionOp(red->definition())) {
-          auto outs = red->definition()->outputs();
-          for (auto out_tv : ir_utils::filterByType<TensorView>(outs)) {
-            reduction_tv_set.insert(out_tv);
-          }
-        }
-      }
-    }
-
-    // If reductions are on fastest dim, don't fuse any operations (after
-    // reductions) that requires an input that is not an input to the
-    // reductions.
-    if (fastest_dim_reduction) {
-      auto post_reduction_vals = DependencyCheck::getAllValsBetween(
-          reduction_tv_set,
-          {fusion->outputs().begin(), fusion->outputs().end()});
-
-      if (post_reduction_vals.empty()) {
-        return true;
-      }
-
-      auto reduction_inputs = IterVisitor::getInputsTo(
-          {reduction_tvs.begin(), reduction_tvs.end()});
-
-      for (auto tv : ir_utils::filterByType<TensorView>(
-               post_reduction_vals.begin(), post_reduction_vals.end())) {
-        if (tv->definition() == nullptr) {
-          continue;
-        }
-
-        auto tv_inputs = IterVisitor::getInputsTo({tv});
-
-        if (std::any_of(
-                tv_inputs.begin(),
-                tv_inputs.end(),
-                [&reduction_inputs](Val* inp) {
-                  return inp->isA<TensorView>() &&
-                      std::find(
-                          reduction_inputs.begin(),
-                          reduction_inputs.end(),
-                          inp) == reduction_inputs.end();
-                })) {
-          return false;
-        }
-      }
-    }
-
-    return true;
-  }
-
-  /* Returns if any non-trivial views are not before the reference. For example:
-   *     t0
-   *    /  \
-   *  view ref
-   *   |
-   *   t1
-   * This could be important as transform propagation from a reference backwards
-   * through a view should always work, but transform propagation form a
-   * reference forward through a view could interfere with the view transforms.
-   */
-  static bool hasViewNotBeforeRef(
-      Fusion* fusion,
-      const std::vector<TensorView*>& reference_tvs) {
-    std::vector<TensorView*> view_tvs;
-    auto view_ops = ir_utils::getViewOps(fusion);
-    for (auto view_op : view_ops) {
-      auto tv_outs = ir_utils::filterByType<TensorView>(view_op->outputs());
-      for (auto entry : tv_outs) {
-        view_tvs.push_back(entry);
-      }
-    }
-
-    if (view_tvs.empty()) {
-      return false;
-    }
-
-    // Terrible complexity, may be worth improving, but is a compile time
-    // check.
-    for (auto ref_tv : reference_tvs) {
-      for (auto view_tv : view_tvs) {
-        if (!DependencyCheck::isDependencyOf(view_tv, ref_tv)) {
-          return true;
-        }
-      }
-    }
-
-    return false;
-  }
-
-  // Checks if there's any gather-like ops that result in non-resolved
-  // broadcast domains and then get squeezed before reaching reduction
-  // TVs. The reduction scheduler uses reduction TVs as a scheduling
-  // reference, so that won't be able to schedule the broadcast ID if
-  // squeezed and its corresponding index-accessed producer ID, and
-  // any IDs that the producer ID depends on.
-  //
-  // This analysis has some similarity as DomainMap. Can be
-  // consolidated?
-  static bool hasGatherToBroadcastBeforeReduction(
-      Fusion* fusion,
-      const std::vector<TensorView*>& reduction_tvs) {
-    std::vector<Val*> reduction_inputs;
-    const auto all_exprs = DependencyCheck::getAllExprsBetween(
-        {fusion->inputs().begin(), fusion->inputs().end()},
-        {reduction_tvs.begin(), reduction_tvs.end()});
-
-    // Grab all consumer domains of indexed domains by gather-like ops
-    std::unordered_set<IterDomain*> broadcast_consumer_of_indexed_ids;
-    for (auto expr : all_exprs) {
-      auto in_tv = ir_utils::getTvInput(expr);
-      // Fusion input does not conflict
-      if (in_tv == nullptr || in_tv->isFusionInput()) {
-        continue;
-      }
-      // In the case of select, there's no consumer domain, and thus
-      // there's no way to schedule the indexed producer domain
-      if (expr->isA<SelectOp>()) {
-        return true;
-      }
-      if (auto consumer_of_indexed_producer =
-              ir_utils::getConsumerOfIndexedProducerID(expr)) {
-        if (consumer_of_indexed_producer->isBroadcast()) {
-          broadcast_consumer_of_indexed_ids.insert(
-              consumer_of_indexed_producer);
-        }
-      }
-    }
-
-    if (broadcast_consumer_of_indexed_ids.empty()) {
-      return false;
-    }
-
-    // If the broadcast IDs are mapped with the reduction TVs, the
-    // reduction scheduler should be able to schedule the gather
-    // output TVs. This mapping can be PERMISSIVE as the broadcast IDs
-    // may be concretized. ExactRootDomainMap may be enough as
-    // broadcasts should not be removed by rfactor exprs.
-
-    // Consider reusing a CA map
-    ComputeAtMap ca_map(fusion);
-    // All of reduction TVs are mapped, so doesn't matter which
-    // reduction tv to use
-    auto ref_tv = reduction_tvs.at(0);
-    return std::any_of(
-        broadcast_consumer_of_indexed_ids.begin(),
-        broadcast_consumer_of_indexed_ids.end(),
-        [&ca_map, &ref_tv](IterDomain* broadcast_consumer_id) {
-          // Check if this broadcast ID has no mapping
-          // with the reference TV.
-          return std::none_of(
-              ref_tv->getRootDomain().begin(),
-              ref_tv->getRootDomain().end(),
-              [&](IterDomain* red_tv_root_id) {
-                return ca_map.areMapped(
-                    broadcast_consumer_id,
-                    red_tv_root_id,
-                    IdMappingMode::PERMISSIVE);
-              });
-        });
-  }
-};
-
-bool isConnectedFusionGraph(Fusion* fusion) {
-  if (fusion->outputs().empty()) {
-    // Trivial case interpreted as connected
-    return true;
-  }
-
-  // A set of connected components on the fusion graph
-  DisjointSets<Val*> component_sets;
-
-  NVF_ERROR(
-      !fusion->outputs().empty(), "Fusion without output is not supported");
-  auto output0 = fusion->outputs()[0];
-  component_sets.initializeSet(output0);
-
-  // Iterate through all used exprs
-  for (auto expr : fusion->exprs()) {
-    NVF_ERROR(!expr->outputs().empty(), "unknown expr with zero output");
-
-    // Each expr maps all its inputs and
-    //  outputs to the same component
-    auto output0 = expr->output(0);
-    for (auto input : ir_utils::filterByType<TensorView>(expr->inputs())) {
-      component_sets.mapEntries(output0, input);
-    }
-    for (auto output : expr->outputs()) {
-      component_sets.mapEntries(output0, output);
-    }
-  }
-
-  // Map aliased outputs
-  for (auto alias_it : fusion->ioAlias()) {
-    component_sets.mapEntries(alias_it.first, alias_it.second);
-  }
-
-  // Check connected-ness:
-  //  If there is no independent compute flow
-  // on this fusion graph, all outputs will be
-  // equivalent/connected to the first output.
-  for (auto output : fusion->outputs()) {
-    if (!component_sets.strictAreMapped(output0, output)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-// Returns if a fusion cannot transformed into a consistent format since we
-// can't transform forward through view operations, for exmaple:
-//
-// tv0[I0, I1, I2]
-// tv1[I0*I1, I2] = view(tv0)
-// tv2[I0, I1*I2] = view(tv0)
-//
-// If we start transform propagation at either tv1 or tv2, it would require
-// "replaying forward" through the other. If we started at tv1 we'd have to be
-// able to take tv2[I0, I1*I2] and transform it to [I0*I1, I2], however this
-// would "undo" the view transformation which we do not support today.
-//
-// Returns true if a scenario like above is found in the fusion.
-bool requiresForwardViewReplay(Fusion* fusion, ComputeAtMap& ca_map) {
-  // Track the uses of the rfactor domains in the fusion. If an rfactor domain
-  // is used in more than one way it means the above situation is being
-  // encountered.
-  //
-  // tv1 root: [I0rf, I1rf, I2] -> rfactor [I0*I1rf, I2]
-  // tv1 root: [I0, I1rf, I2rf] -> rfactor [I0, I1*I2rf]
-  //
-  // Here we can see I1rf is used in two view transformations, one to I0*I1rf,
-  // and the other to I1*I2rf.
-
-  // Track the transformation each exact disjoint rfactor set is used in. If
-  // more than one is detected we can't support transforming the fusion into a
-  // consistent format.
-  std::unordered_map<std::shared_ptr<VectorOfUniqueEntries<IterDomain*>>, Expr*>
-      unique_exact_uses;
-
-  // Don't check compute uses directly, as IterDomain->uses() isn't protected
-  // from going outside the TensorViews between registered inputs and outputs of
-  // the fusion. If there are view operations defined in the fusion container
-  // (because of how segmentation works) but not between registered input and
-  // outputs, that could be picked up as inconsistent view transformations.
-  //
-  // It would be unlikely this would be picked up as a conflict as we check
-  // which definitions were registered in the compute at map for matching
-  // transformations. However, we may want to support scheduling after
-  // transformations which could map to those views not on the input->output
-  // path.
-
-  // Look through all definitions associated with producing rfactor outputs.
-  // Mark those as an active use of the rfactor, if two are detected, return
-  // true.
-  for (const auto& disjoint_set_shared_ptr :
-       ca_map.idGraph().exactNodes().disjointSets()) {
-    // Make sure there's at least one rfactor domain in the set, otherwise we
-    // don't need to check anything from this set.
-    if (!std::any_of(
-            disjoint_set_shared_ptr->vector().begin(),
-            disjoint_set_shared_ptr->vector().end(),
-            [](IterDomain* id) { return id->isRFactorProduct(); })) {
-      continue;
-    }
-
-    // Grab all the unique definitions detected to consume the iter domains in
-    // this set
-    auto unique_defs =
-        ca_map.uniqueExactDefinitions(disjoint_set_shared_ptr->back());
-
-    // Iterate through the all the rfactor iter domains
-    for (auto id_rfactor_product : disjoint_set_shared_ptr->vector()) {
-      if (!id_rfactor_product->isRFactorProduct()) {
-        continue;
-      }
-
-      // Grab the rfactor definition
-      auto rfactor_def = id_rfactor_product->definition();
-
-      if (rfactor_def == nullptr) {
-        // Guard segfault if there isn't a definition for this iter domain
-        continue;
-      }
-
-      // rfactor_def can be Resize, but resize transformation is not
-      // replayed, so mismatch doesn't matter
-      if (rfactor_def->isA<Resize>()) {
-        continue;
-      }
-
-      // If one output of the expression is an rfactor ID all of them should be
-      auto def_outs =
-          ir_utils::filterByType<IterDomain>(rfactor_def->outputs());
-      NVF_ERROR(
-          std::all_of(
-              def_outs.begin(),
-              def_outs.end(),
-              [](IterDomain* id) { return id->isRFactorProduct(); }),
-          "This function does not support outputs of transformations with mismatching rfactor flags. ",
-          "If one output is rfactor all should be rfactor.");
-
-      // If outputs are rfactor all the inputs should be as well. It doesn't
-      // make sense to have transforms on non-rfactor domains that produce
-      // rfactor domains.
-      auto def_inps = ir_utils::filterByType<IterDomain>(rfactor_def->inputs());
-      NVF_ERROR(
-          std::all_of(
-              def_inps.begin(),
-              def_inps.end(),
-              [](IterDomain* id) { return id->isRFactorProduct(); }),
-          "Inputs producing an rfactor domain, should be marked as rfactor but found:\n  ",
-          rfactor_def->toString());
-
-      // Check which definition in the unique exact definition set this
-      // definition matches to:
-      // TODO: Why does it need to check all unique defs? It actually
-      // only looks at those that are exact with rfactor_def and
-      // adds those unique_defs, which are all exactly mapped, to the
-      // unique_exact_use map. Since the objective of this analysis is
-      // to find non-exact exprs using the same exact ID set, it seems
-      // it's just sufficient to register rfactor_def as a user of the
-      // exact set.
-      for (auto unique_def : unique_defs) {
-        if (ca_map.areExactExprs(rfactor_def, unique_def)) {
-          // Check if we already have an expression that consumes an
-          // equivalent of any of the input rfactor domains. If so and it's
-          // not the already registered transformation, return true
-          for (auto inp : def_inps) {
-            auto inp_disjoint_set =
-                ca_map.disjointSetOf(inp, IdMappingMode::EXACT);
-            // Initialize the use entry for this set (if it doesn't already
-            // exist)
-            if (unique_exact_uses.find(inp_disjoint_set) ==
-                unique_exact_uses.end()) {
-              unique_exact_uses[inp_disjoint_set] = nullptr;
-            }
-
-            if (unique_exact_uses.at(inp_disjoint_set) == nullptr) {
-              // If expression is null pointer register this unique_def
-              unique_exact_uses[inp_disjoint_set] = unique_def;
-            } else if (!ca_map.areExactExprs(
-                           unique_exact_uses[inp_disjoint_set], unique_def)) {
-              // Two transformations that don't match on matching rfactor
-              // domains found, return true.
-              return true;
-            }
-          }
-          // Expression already mapped, stop trying to match expressions
-          break;
-        }
-      }
-    }
-  }
-  // No inconsistent rfactor uses found, we can safely transform this graph.
-  return false;
-}
-
-// Returns if view interferes with how we want to treat the reference, being at
-// least a 2D reduction schedule but maybe a 3D reduction schedule.
-bool reductionInterferingView(
-    Fusion* fusion,
-    const ComputeAtMap& ca_map,
-    TensorView* reduction_reference) {
-  // Make sure the view doesn't interfere with how we'll want to schedule
-  // it. If we might want to do a 3D scheduler make sure views are disjoint
-  // based on what the 3D scheduler's merges would be.
-
-  // Utility to take dimensions out of the vector that we've already
-  // processed or don't want to process.
-  auto remove_dims = [](const std::vector<IterDomain*>& dims,
-                        std::unordered_set<IterDomain*> to_remove) {
-    std::vector<IterDomain*> dims_removed;
-    std::copy_if(
-        dims.begin(),
-        dims.end(),
-        std::back_inserter(dims_removed),
-        [&](IterDomain* id) { return to_remove.find(id) == to_remove.end(); });
-    return dims_removed;
-  };
-
-  std::vector<IterDomain*> dims = reduction_reference->getMaybeRFactorDomain();
-
-  // The disjoint groups we need for this scheduler
-  std::vector<std::vector<IterDomain*>> groups;
-
-  // Do this three times as we could have a 3D scheduler at maximum
-  for (auto dimension : c10::irange(3)) {
-    // Tracker for this group
-    std::vector<IterDomain*> current_dims;
-
-    // Tracker of what we've already processed to remove from dims
-    std::unordered_set<IterDomain*> processed;
-
-    for (auto i : c10::irange(dims.size())) {
-      auto dim_i = dims.size() - i - 1;
-      if (dims[dim_i]->isReduction() != dims[dims.size() - 1]->isReduction()) {
-        if (dimension == 0) {
-          // First dimension must be contiguous merges
-          break;
-        } else {
-          // Other dimensions can be non contiguous merges
-          continue;
-        }
-      }
-      current_dims.push_back(dims[dim_i]);
-      processed.emplace(dims[dim_i]);
-    }
-
-    // Don't add empty group (would happen if it's a 2D scheduler not 3D)
-    if (!current_dims.empty()) {
-      groups.push_back(current_dims);
-      dims = remove_dims(dims, processed);
-    }
-  }
-
-  NVF_ERROR(dims.empty(), "Error processing ", dims, " in registry.cpp.");
-
-  // Make sure groups are disjoint based on view
-
-  auto disjoint_rfactor_sets = scheduler_utils::disjointRFactorSets(fusion);
-  auto disjoint_set_information = scheduler_utils::getDisjointRFactorSetsOf(
-      fusion, reduction_reference, disjoint_rfactor_sets);
-
-  // Convert id's in groups to disjoint_set_ids of disjoint_set_information
-  std::vector<std::vector<int>> disjoint_groups;
-
-  for (const auto& group : groups) {
-    std::vector<int> disjoint_id_sets;
-    for (auto id : group) {
-      auto find_it = std::find(
-          reduction_reference->getMaybeRFactorDomain().begin(),
-          reduction_reference->getMaybeRFactorDomain().end(),
-          id);
-      NVF_ERROR(
-          find_it != reduction_reference->getMaybeRFactorDomain().end(),
-          "Issue with view analysis on reduction like schedule, with reference: ",
-          reduction_reference->toString());
-      auto rfactor_pos = std::distance(
-          reduction_reference->getMaybeRFactorDomain().begin(), find_it);
-      NVF_ERROR(
-          rfactor_pos < (int)disjoint_set_information.disjoint_set_ids.size(),
-          "Error computing disjoint group on the rfactor domain of ",
-          reduction_reference->toString());
-      disjoint_id_sets.push_back(
-          disjoint_set_information.disjoint_set_ids[rfactor_pos]);
-    }
-    disjoint_groups.push_back(disjoint_id_sets);
-  }
-
-  // Make sure there's no intersection between the groups, otherwise view
-  // will interfere with the schedule. TODO: Make this better complexity,
-  // since it should be relatively small int vectors of a small total nDims,
-  // not too worried about it now.
-
-  for (auto first_dim_i : c10::irange(disjoint_groups.size())) {
-    for (auto second_dim_i = first_dim_i + 1;
-         second_dim_i < disjoint_groups.size();
-         ++second_dim_i) {
-      auto first_group = disjoint_groups[first_dim_i];
-      auto second_group = disjoint_groups[second_dim_i];
-      for (auto first_disjoint_id : first_group) {
-        for (auto second_disjoint_id : second_group) {
-          if (first_disjoint_id == second_disjoint_id) {
-            return true;
-          }
-        }
-      }
-    }
-  }
-  return false;
-}
-
-PrimDataType getTensorIndexType(TensorView* tv, ExpressionEvaluator& ee) {
-  NVF_ERROR(
-      !tv->isFusionInput(),
-      "This function is not supposed to be used for fusion inputs: ",
-      tv->toString());
-
-  auto non_contig = std::any_of(
-      tv->domain()->contiguity().begin(),
-      tv->domain()->contiguity().end(),
-      [](const auto contig) { return contig.has_value() && !contig.value(); });
-
-  // When a fusion output is non-contiguous, currently there's no
-  // way to obtain its strides. This is an interface problem and
-  // should be fixed.
-  if (tv->isFusionOutput() && non_contig) {
-    return PrimDataType::Int;
-  }
-
-  // This function should not be used for fusion inputs, so any
-  // non-contig tensor means a fusion intermediate tensor. However,
-  // since we don't support non-contiguous intermediates, there must be
-  // something wrong.
-  NVF_ERROR(
-      !non_contig, "Unexpected non-contiguous tensor found: ", tv->toString());
-
-  // Note that at this point tensors are not scheduled yet. Each
-  // tensor may end up being inlined, stored on Shared or Local, but
-  // the index type is currently supposed to be determined before
-  // any of scheduling decisions is made, so we may end up making a
-  // conservative decision.
-  // TODO: Consider index type resolution after segmentation and
-  // scheduling. At that point we have the final scheduled fusions
-  // with which we can make more precise analyses. It would require
-  // that the scheduling and segmentation should not have any
-  // assumption about the index type as it may change.
-  int64_t stride = 1;
-  KernelIndexTypeCompute index_type_helper;
-  for (auto i = tv->getMaybeRFactorDomain().size(); i > 0; --i) {
-    auto id = tv->getMaybeRFactorDomain().at(i - 1);
-    if (id->isReduction() || id->isStride() || id->isBroadcast()) {
-      continue;
-    }
-
-    auto extent = ee.evaluate(id->extent());
-    // We could also just conservatively use 64-bit indexing if the
-    // extent size is not determined, but this should be possible to
-    // evaluate.
-    NVF_ERROR(
-        extent.hasValue(),
-        "Axis with unknown extent found: ",
-        id->toString(),
-        ", tensor: ",
-        tv->toString());
-
-    auto extent_int = extent.as<int64_t>();
-
-    NVF_ERROR(extent_int >= 0, "Unexpected size of axis: ", extent_int);
-
-    if (extent_int > 0) {
-      if (index_type_helper.addDim(extent.as<int64_t>(), stride) ==
-          PrimDataType::Int) {
-        return PrimDataType::Int;
-      }
-      stride *= extent.as<int64_t>();
-    }
-  }
-
-  return index_type_helper.getType();
-}
-
-// Check inputs, outputs and intermediates
-// Intermediates are contiguous, so strides are not necessary
-// Strides are required for inputs and also maybe for outputs as
-// they may be non-contiguous. However, in our current interface,
-// output strides are not available, so if there's any outputs that
-// are non contiguous, need to fall back to 64-bit indexing
-PrimDataType getIndexTypeOfKernel(
-    Fusion* fusion,
-    const std::vector<TensorView*>& all_tvs,
-    const KernelArgumentHolder& inputs,
-    ExpressionEvaluator& ee) {
-  if (inputs.getSmallestIndexTypeOfArguments() == PrimDataType::Int) {
-    return PrimDataType::Int;
-  }
-
-  for (auto tv : all_tvs) {
-    // Fusion input tensors are included in the args parameter, and
-    // they are checked separately
-    if (tv->isFusionInput()) {
-      continue;
-    }
-
-    if (getTensorIndexType(tv, ee) == PrimDataType::Int) {
-      return PrimDataType::Int;
-    }
-  }
-
-  return PrimDataType::Int32;
-}
-
-} // namespace
 
 SchedulerRuntimeInfo::SchedulerRuntimeInfo(
     Fusion* complete_fusion,
@@ -952,7 +45,7 @@ SchedulerRuntimeInfo::SchedulerRuntimeInfo(
   if (forced_index_type.has_value()) {
     index_type_ = forced_index_type.value();
   } else {
-    index_type_ = getIndexTypeOfKernel(
+    index_type_ = registry_utils::getIndexTypeOfKernel(
         complete_fusion_,
         all_tvs.empty() ? ir_utils::allTvs(complete_fusion_) : all_tvs,
         args,
@@ -1058,62 +151,6 @@ bool SchedulerEntry::sameAs(const SchedulerEntry* other) {
 
 namespace {
 
-static bool checkPatternEquivalence(
-    TensorView* out_tv0,
-    TensorView* out_tv1,
-    const ComputeAtRootDomainMap& root_map) {
-  const auto& out_root0 = out_tv0->getRootDomain();
-  const auto& out_root1 = out_tv1->getRootDomain();
-  const auto domain0 = out_tv0->domain();
-  const auto domain1 = out_tv1->domain();
-
-  auto it0 = out_root0.begin();
-  auto it1 = out_root1.begin();
-
-  auto skip_broadcast = [&]() {
-    while (it0 != out_root0.end() && (*it0)->isBroadcast()) {
-      it0++;
-    }
-    while (it1 != out_root1.end() && (*it1)->isBroadcast()) {
-      it1++;
-    }
-  };
-
-  skip_broadcast();
-  while (it0 != out_root0.end() && it1 != out_root1.end()) {
-    if ((*it0)->isReduction() != (*it1)->isReduction()) {
-      return false;
-    }
-    if (!root_map.canMap(domain0, (*it0), domain1, (*it1))) {
-      return false;
-    }
-    it0++;
-    it1++;
-    skip_broadcast();
-  }
-
-  return it0 == out_root0.end() && it1 == out_root1.end();
-}
-
-// Reusing some code from lowering specifically in lower_trivial_broadcast.cpp
-// ConcretizedBroadcastDomains::maybeNonUniquelyConcretized this checks if
-// there's a broadcast iteration domain that's being broadcasted to seemingly
-// different extents, meaning we don't know in the kernel if the dimension is
-// being broadcasted to one size multiple times or different sizes. This is a
-// hard to optimize problem and likely indicates we shouldn't be fusing.
-bool hasNonUniqueBcast(Fusion* fusion) {
-  ConcretizedBroadcastDomains concretize_info(fusion);
-
-  for (auto tv : ir_utils::allTvs(fusion)) {
-    for (auto id : tv->getRootDomain()) {
-      if (concretize_info.maybeNonUniquelyConcretized(id)) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
 //! Scheduler interface:
 //!    Each of the scheduler needs to provide 3 interface functions:
 //!
@@ -1213,7 +250,8 @@ class NoOpScheduler : public SchedulerEntry {
     }
 
     // Check that inputs of all select/gather-like ops are fusion inputs
-    if (rejectScheduleForMemoryPromotion(fusion, ScheduleHeuristic::NoOp)) {
+    if (registry_utils::rejectScheduleForMemoryPromotion(
+            fusion, ScheduleHeuristic::NoOp)) {
       return false;
     }
 
@@ -1275,7 +313,7 @@ class ReductionScheduler : public SchedulerEntry {
     }
 
     // Check that inputs of all select/gather-like ops are fusion inputs
-    if (rejectScheduleForMemoryPromotion(
+    if (registry_utils::rejectScheduleForMemoryPromotion(
             fusion, ScheduleHeuristic::Reduction)) {
       return false;
     }
@@ -1294,7 +332,7 @@ class ReductionScheduler : public SchedulerEntry {
       return false;
     }
 
-    if (hasNonUniqueBcast(fusion)) {
+    if (registry_utils::hasNonUniqueBcast(fusion)) {
       scheduler_debug_utils::canScheduleRejectReason(
           ScheduleHeuristic::Reduction,
           "Broadcasting dimension might be broadcasting to multiple sizes.");
@@ -1303,7 +341,7 @@ class ReductionScheduler : public SchedulerEntry {
 
     if (!ir_utils::getViewOps(fusion).empty()) {
       ComputeAtMap ca_map(fusion);
-      if (requiresForwardViewReplay(fusion, ca_map)) {
+      if (registry_utils::requiresForwardViewReplay(fusion, ca_map)) {
         scheduler_debug_utils::canScheduleRejectReason(
             ScheduleHeuristic::Reduction,
             "Fusion requires view being reversible.");
@@ -1312,7 +350,8 @@ class ReductionScheduler : public SchedulerEntry {
 
       // Reduction scheduler simply uses reduction_tvs[0] as the reference, if
       // that changes, this needs to be changed.
-      if (reductionInterferingView(fusion, ca_map, reduction_tvs[0])) {
+      if (registry_utils::reductionInterferingView(
+              fusion, ca_map, reduction_tvs[0])) {
         scheduler_debug_utils::canScheduleRejectReason(
             ScheduleHeuristic::Reduction,
             "View may interfere with reduction scheduling.");
@@ -1362,7 +401,7 @@ class ReductionScheduler : public SchedulerEntry {
 
       // red_ops.size()>1 checked before
       for (size_t it = 1; it < reduction_tvs.size(); it++) {
-        if (!checkPatternEquivalence(
+        if (!registry_utils::checkPatternEquivalence(
                 reduction_tvs[it - 1], reduction_tvs[it], root_map)) {
           scheduler_debug_utils::canScheduleRejectReason(
               ScheduleHeuristic::Reduction,
@@ -1384,17 +423,18 @@ class ReductionScheduler : public SchedulerEntry {
       return false;
     }
 
-    if (!SchedulerTopologyChecker::supportedPostReductionFusion(
+    if (!registry_utils::SchedulerTopologyChecker::supportedPostReductionFusion(
             fusion, reduction_tvs) ||
-        SchedulerTopologyChecker::hasPostReductionBCast(fusion)) {
+        registry_utils::SchedulerTopologyChecker::hasPostReductionBCast(
+            fusion)) {
       scheduler_debug_utils::canScheduleRejectReason(
           ScheduleHeuristic::Reduction,
           "has unsupported post reduction fusion");
       return false;
     }
 
-    if (SchedulerTopologyChecker::hasGatherToBroadcastBeforeReduction(
-            fusion, reduction_tvs)) {
+    if (registry_utils::SchedulerTopologyChecker::
+            hasGatherToBroadcastBeforeReduction(fusion, reduction_tvs)) {
       scheduler_debug_utils::canScheduleRejectReason(
           ScheduleHeuristic::Reduction,
           "has unsupported gather-like ops before reduction");
@@ -1438,7 +478,7 @@ class TransposeScheduler : public SchedulerEntry {
 
   static bool canScheduleCompileTime(Fusion* fusion) {
     // Check that inputs of all select/gather-like ops are fusion inputs
-    if (rejectScheduleForMemoryPromotion(
+    if (registry_utils::rejectScheduleForMemoryPromotion(
             fusion, ScheduleHeuristic::Transpose)) {
       return false;
     }
@@ -1499,7 +539,7 @@ class TransposeScheduler : public SchedulerEntry {
       return false;
     }
 
-    if (hasNonUniqueBcast(fusion)) {
+    if (registry_utils::hasNonUniqueBcast(fusion)) {
       scheduler_debug_utils::canScheduleRejectReason(
           ScheduleHeuristic::Transpose,
           "Broadcasting dimension might be broadcasting to multiple sizes.");
@@ -1561,7 +601,7 @@ class PointWiseScheduler : public SchedulerEntry {
     }
 
     // Check that inputs of all select/gather-like ops are fusion inputs
-    if (rejectScheduleForMemoryPromotion(
+    if (registry_utils::rejectScheduleForMemoryPromotion(
             fusion, ScheduleHeuristic::PointWise)) {
       return false;
     }
@@ -1575,7 +615,7 @@ class PointWiseScheduler : public SchedulerEntry {
 
     if (!ir_utils::getViewOps(fusion).empty()) {
       ComputeAtMap ca_map(fusion);
-      if (requiresForwardViewReplay(fusion, ca_map)) {
+      if (registry_utils::requiresForwardViewReplay(fusion, ca_map)) {
         scheduler_debug_utils::canScheduleRejectReason(
             ScheduleHeuristic::PointWise,
             "Fusion requires view being reversible.");
@@ -1591,7 +631,7 @@ class PointWiseScheduler : public SchedulerEntry {
       return false;
     }
 
-    if (hasNonUniqueBcast(fusion)) {
+    if (registry_utils::hasNonUniqueBcast(fusion)) {
       scheduler_debug_utils::canScheduleRejectReason(
           ScheduleHeuristic::PointWise,
           "Broadcasting dimension might be broadcasting to multiple sizes.");
@@ -1631,654 +671,6 @@ class PointWiseScheduler : public SchedulerEntry {
       HeuristicSummary* data_cache = nullptr) {
     params_ = getPointwiseHeuristics(fusion, runtime_info, data_cache);
     NVF_ERROR(params_ != nullptr);
-  }
-};
-
-class PersistentKernelScheduler : public SchedulerEntry {
- public:
-  explicit PersistentKernelScheduler(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr)
-      : SchedulerEntry(ScheduleHeuristic::Persistent) {
-    computeHeuristics(fusion, runtime_info, data_cache);
-  }
-
-  void schedule(Fusion* fusion) override {
-    FUSER_PERF_SCOPE("Schedule Persistent Fusion");
-    schedulePersistentKernel(fusion, reductionParams());
-  }
-
-  static bool canScheduleCompileTime(Fusion* fusion) {
-    // Needs at least one reduction to consider.
-    auto reduction_ops = ir_utils::getReductionOps(fusion);
-    if (reduction_ops.empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "needs a reduction op");
-      return false;
-    }
-
-    if (ir_utils::filterByType<TensorView>(fusion->inputs()).empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "Scheduling not supported with no input");
-      return false;
-    }
-
-    // Check that inputs of all select/gather-like ops are fusion inputs
-    if (rejectScheduleForMemoryPromotion(
-            fusion, ScheduleHeuristic::Persistent)) {
-      return false;
-    }
-
-    // Fusions handled by persistent kernel scheduler cannot have MmaOp.
-    if (!ir_utils::getMmaOps(fusion).empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "no support for mma ops.");
-      return false;
-    }
-
-    if (hasNonUniqueBcast(fusion)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "Broadcasting dimension might be broadcasting to multiple sizes.");
-      return false;
-    }
-
-    auto reduction_tvs = scheduler_utils::getReductionTvs(fusion);
-
-    if (reduction_tvs.empty()) {
-      // Use pointwise logic
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "no reduction tv");
-      return false;
-    }
-
-    std::vector<TensorView*> inner_reduction_tvs;
-    std::vector<TensorView*> outer_reduction_tvs;
-    for (auto tv : reduction_tvs) {
-      if (scheduler_utils::isFastestDimReduction(tv)) {
-        inner_reduction_tvs.emplace_back(tv);
-      } else {
-        outer_reduction_tvs.emplace_back(tv);
-      }
-    }
-    bool combined_inner_outer =
-        !inner_reduction_tvs.empty() && !outer_reduction_tvs.empty();
-    if (!checkReductionPattern(
-            fusion, inner_reduction_tvs, outer_reduction_tvs)) {
-      return false;
-    }
-    // If there is both inner and outer reduction, we use the first inner
-    // reduction tv as reference, otherwise we use the first reduction tv,
-    // whether it is inner or outer.
-    TensorView* reference_tv =
-        combined_inner_outer ? inner_reduction_tvs[0] : reduction_tvs[0];
-
-    if (!ir_utils::getViewOps(fusion).empty()) {
-      ComputeAtMap ca_map(fusion);
-      if (requiresForwardViewReplay(fusion, ca_map)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "Fusion requires view being reversible.");
-        return false;
-      }
-
-      // Persistent scheduler simply uses reference_tv as the reference, if
-      // that changes, this needs to be changed.
-      if (reductionInterferingView(fusion, ca_map, reference_tv)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "View may interfere with normalization scheduling.");
-        return false;
-      }
-    }
-
-    // Before examining the reduction axes want to quickly
-    //   check the reductions have the same axis width
-    //   to avoid building root domain map in easier cases
-    bool valid_axis_count = false;
-    size_t axis_count = 0;
-    auto reduction_root_size = [](TensorView* red_tv) {
-      size_t count = 0;
-      for (auto id : red_tv->getRootDomain()) {
-        if (!id->isBroadcast()) {
-          count++;
-        }
-      }
-      return count;
-    };
-
-    for (auto red : reduction_tvs) {
-      if (!valid_axis_count) {
-        valid_axis_count = true;
-        axis_count = reduction_root_size(red);
-      } else {
-        if (reduction_root_size(red) != axis_count) {
-          scheduler_debug_utils::canScheduleRejectReason(
-              ScheduleHeuristic::Persistent,
-              "inconsistent reduction root size: ",
-              red->toString(),
-              ", expected: ",
-              axis_count);
-          return false;
-        }
-      }
-    }
-
-    // Only accept persistent kernels
-    auto persistent_buffer_info = scheduler_utils::persistentBuffers(fusion);
-    if (persistent_buffer_info.persistent_buffers.empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "no persistent buffer identified");
-      return false;
-    }
-
-    if (SchedulerTopologyChecker::hasNonNormalizePostReductionBCast(fusion)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "unsupported post reduction normalization");
-      return false;
-    }
-
-    if (SchedulerTopologyChecker::hasGatherToBroadcastBeforeReduction(
-            fusion, reduction_tvs)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "has unsupported gather-like ops before normalization");
-      return false;
-    }
-
-    return true;
-  }
-
-  static bool canScheduleRunTime(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    FUSER_PERF_SCOPE("PersistentKernelScheduler::canSchedule");
-    auto reduction_tv_entry =
-        HeuristicSummaryEntry<HeuristicCompileTime::ReductionTVs>(
-            data_cache, [&fusion]() {
-              return std::make_unique<std::vector<TensorView*>>(
-                  scheduler_utils::getReductionTvs(fusion));
-            });
-
-    auto& reduction_tvs = reduction_tv_entry.get();
-    bool inner_reduction = false;
-    bool outer_reduction = false;
-    TensorView* first_inner_reduction_tv = nullptr;
-    for (auto tv : reduction_tvs) {
-      if (scheduler_utils::isFastestDimReduction(tv)) {
-        first_inner_reduction_tv = tv;
-        inner_reduction = true;
-      } else {
-        outer_reduction = true;
-      }
-    }
-
-    // If there is both inner and outer reduction, we use the first inner
-    // reduction tv to get properties, otherwise we use the first reduction tv,
-    // whether it is inner or outer.
-    auto reference_tv = inner_reduction && outer_reduction
-        ? first_inner_reduction_tv
-        : reduction_tvs[0];
-
-    auto properties = scheduler_utils::getReductionProperties(
-        fusion, runtime_info, reference_tv);
-
-    const int64_t warp_size = at::cuda::getCurrentDeviceProperties()->warpSize;
-
-    if (!properties.fastest_dim_reduction) {
-      return canScheduleRunTimeOuter(
-          fusion, runtime_info, data_cache, reduction_tvs, properties);
-    }
-
-    // pair of persistent_buffer_size and available_persistent_buffer_size
-    const std::pair<int64_t, int64_t> buffer_size = getPersistentBufferSize(
-        fusion, runtime_info, data_cache, reduction_tvs);
-    const int64_t persistent_buffer_size = buffer_size.first;
-    const int64_t available_persistent_buffer_size = buffer_size.second;
-
-    const int64_t device_multiprocessor_count =
-        (int64_t)at::cuda::getCurrentDeviceProperties()->multiProcessorCount;
-
-    if (persistent_buffer_size > available_persistent_buffer_size) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "not enough registers or shared memory for persistence");
-      return false;
-    }
-
-    if (inner_reduction && outer_reduction) {
-      // get vectorize_factor, same process to that in getPersistentHeuristics
-      auto reduced_tv = ir_utils::getSoleProducerTv(reference_tv);
-      const auto vectorize_factor = vectorize_helper::getVectorizationFactor(
-          runtime_info,
-          reduced_tv,
-          data_cache,
-          (int)(reduced_tv->nDims() - properties.inner_most_dimension_ndims));
-      // check if we can schedule the combined reductions with a reasonable
-      // batch size without register spills.
-      if (!normalization_scheduler_utils::
-               getOptionalInnerOuterPersistentBufferBatches(
-                   properties.total_reduction_numel,
-                   properties.total_iteration_numel,
-                   persistent_buffer_size,
-                   (int64_t)vectorize_factor,
-                   warp_size,
-                   false)
-                   .first.has_value()) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "Required batch number is larger than available batch number! Will cause register spills!");
-        return false;
-      }
-    }
-
-    const int64_t device_max_threads_per_multiprocessor =
-        (int64_t)at::cuda::getCurrentDeviceProperties()
-            ->maxThreadsPerMultiProcessor;
-
-    // Maximum number of iteration dimensions we can have and still be
-    // persistent.
-    const int64_t max_multi_reduction_factor = scheduler_utils::safeDiv(
-        available_persistent_buffer_size, persistent_buffer_size);
-
-    const int64_t required_sm_per_norm =
-        ceilDiv(persistent_buffer_size, scheduler_utils::register_file_size);
-
-    // If the persistence requires over half the device don't do grid
-    // persistence as we can't overlap the grid comms.
-    if (required_sm_per_norm >
-        scheduler_utils::safeDiv(device_multiprocessor_count, 3)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "requires over half GPU persistence.");
-      return false;
-    }
-
-    const int64_t norm_per_sm =
-        ceilDiv(scheduler_utils::register_file_size, persistent_buffer_size);
-
-    // If outer reduction, don't go persistent if we can't fit half a warp in
-    // the iter domain of the persistent reduction.
-    if (!properties.fastest_dim_reduction &&
-        !(norm_per_sm >= warp_size / 2 ||
-          max_multi_reduction_factor >= warp_size)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "not enough threads");
-      return false;
-    }
-
-    // Don't go persistent if we can't use a small fraction of the
-    // available SMs yet have a large reduction size.
-    if ( // Large reduction dim
-        properties.total_reduction_numel >=
-            device_max_threads_per_multiprocessor * 4 &&
-        properties.total_iteration_numel <
-            (properties.fastest_dim_reduction
-                 ? scheduler_utils::safeDiv(device_multiprocessor_count, 8)
-                 // Make sure we at least use a quarter of the device * a
-                 // half warp
-                 : (warp_size / 8) * device_multiprocessor_count)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "not enough blocks");
-      return false;
-    }
-
-    return true;
-  }
-
- private:
-  void computeHeuristics(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    params_ = getPersistentHeuristics(fusion, runtime_info, data_cache);
-    NVF_ERROR(params_ != nullptr);
-  }
-
-  static bool checkReductionPattern(
-      Fusion* fusion,
-      const std::vector<TensorView*>& inner_reduction_tvs,
-      const std::vector<TensorView*>& outer_reduction_tvs) {
-    // Use root domain map to check the reduction ops have the same axes
-    FusionGuard fg(fusion);
-    ComputeAtRootDomainMap root_map;
-    root_map.build(true);
-
-    // check inner and outer reductions seperately
-    for (const auto& rtvs : {inner_reduction_tvs, outer_reduction_tvs}) {
-      for (const auto it : c10::irange(1, rtvs.size())) {
-        if (!checkPatternEquivalence(rtvs[it - 1], rtvs[it], root_map)) {
-          scheduler_debug_utils::canScheduleRejectReason(
-              ScheduleHeuristic::Persistent,
-              "unmapped reduction ",
-              rtvs[it - 1],
-              " and ",
-              rtvs[it]);
-          return false;
-        }
-      }
-    }
-    // combined inner and outer reduction is of general purpose but only tested
-    // for layer norm backward
-    if (!inner_reduction_tvs.empty() && !outer_reduction_tvs.empty()) {
-      if (!normalization_scheduler_utils::checkIfReductionsAreInnerOuter(
-              inner_reduction_tvs, outer_reduction_tvs)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "to use combined reduction, inner reduction tensor should be [I,I,...,R,R] and outer reduction tensor should be [R,R,...,I,I]");
-        return false;
-      }
-
-      if (!normalization_scheduler_utils::hasSharedInput(
-              inner_reduction_tvs, outer_reduction_tvs)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "to use combined reduction, inner reduction and outer reduction should have shared input.");
-        return false;
-      }
-
-      if (!normalization_scheduler_utils::
-              isConnectedOnlyThroughReductionProducer(
-                  inner_reduction_tvs, outer_reduction_tvs)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent,
-            "to use combined reduction, inner reduction and outer reduction should not have shared consumer, their consumers should not have shared non-outer-reduction producer.");
-        return false;
-      }
-    }
-    return true;
-  }
-
-  static std::pair<int64_t, int64_t> getPersistentBufferSize(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache,
-      const std::vector<TensorView*>& reduction_tvs) {
-    auto persistent_buffer_info_entry =
-        HeuristicSummaryEntry<HeuristicCompileTime::PersistentBufferInfo>(
-            data_cache, [&fusion]() {
-              return std::make_unique<scheduler_utils::PersistentBufferInfo>(
-                  scheduler_utils::persistentBuffers(fusion));
-            });
-
-    auto& persistent_buffer_info = persistent_buffer_info_entry.get();
-
-    auto persistent_buffer_size_info = scheduler_utils::persistentBufferSize(
-        fusion, runtime_info, persistent_buffer_info, data_cache);
-
-    // Note that projected buffer size can be zero
-    auto persistent_buffer_size =
-        persistent_buffer_size_info.projected_persistent_buffer_size == 0
-        ? persistent_buffer_size_info.persistent_buffer_size
-        : std::min(
-              persistent_buffer_size_info.persistent_buffer_size,
-              persistent_buffer_size_info.projected_persistent_buffer_size);
-
-    // in combined_inner_outer_reduction, the partial results of outer
-    // reductions must be persistent, allow register spill avoid segmentation
-    int64_t inner_reduction_count = 0;
-    int64_t outer_reduction_count = 0;
-    std::vector<TensorView*> outer_reduction_tvs;
-    for (auto tv : reduction_tvs) {
-      if (scheduler_utils::isFastestDimReduction(tv)) {
-        inner_reduction_count++;
-      } else {
-        outer_reduction_count++;
-        outer_reduction_tvs.emplace_back(tv);
-      }
-    }
-    const bool combined_inner_outer_reduction =
-        inner_reduction_count && outer_reduction_count;
-    if (combined_inner_outer_reduction) {
-      persistent_buffer_size +=
-          normalization_scheduler_utils::partialReductionBufferSize(
-              outer_reduction_tvs, runtime_info);
-    }
-
-    // At this point, we use the full register file size only for the
-    // inner-outer case. It does not mean the full size shouldn't be used
-    // otherwise, but more detailed tuning of the heuristics would be required.
-    int64_t available_persistent_buffer_size = combined_inner_outer_reduction
-        ? scheduler_utils::register_file_size_full
-        : scheduler_utils::register_file_size;
-
-    // Use shared memory for persistent buffer is only tested for inner
-    // reduction
-    // TODO: extend to outer reduction and combined reduction
-    const bool allow_shared_memory =
-        inner_reduction_count > 0 && outer_reduction_count == 0;
-    if (allow_shared_memory) {
-      const auto dev_prop = at::cuda::getCurrentDeviceProperties();
-      const int64_t max_shared_memory_size =
-          (int64_t)dev_prop->sharedMemPerBlockOptin;
-      // Some shared memories are reserved for kernel launch overhead and
-      // reduction_broadcast_workspace. Estimation is conservative, but should
-      // be good enough. The actual threads per block is set in the heuristics
-      // and it may be smaller than maxThreadsPerBlock.
-      // TODO: More accurate estimation of available shared memory size.
-      const int64_t kernel_overhead =
-          (int64_t)dev_prop->reservedSharedMemPerBlock;
-      int64_t max_buffer_dtype_size = 1;
-      for (auto tv : persistent_buffer_info.persistent_buffers) {
-        max_buffer_dtype_size = std::max(
-            max_buffer_dtype_size,
-            dataTypeSize(
-                tv->getDataType().value(), runtime_info.getIndexType()));
-      }
-      const int64_t reduction_broadcast_workspace =
-          (int64_t)(dev_prop->maxThreadsPerBlock) * max_buffer_dtype_size;
-      const int64_t available_shared_memory_size = max_shared_memory_size -
-          kernel_overhead - reduction_broadcast_workspace;
-      available_persistent_buffer_size = std::max(
-          available_persistent_buffer_size, available_shared_memory_size);
-    }
-
-    return std::make_pair(
-        persistent_buffer_size, available_persistent_buffer_size);
-  }
-
-  static bool canScheduleRunTimeOuter(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache,
-      const std::vector<TensorView*>& reduction_tvs,
-      const scheduler_utils::ReductionTvProperties& properties) {
-    FUSER_PERF_SCOPE("PersistentKernelScheduler::canScheduleRuntimeOuter");
-    FusionGuard fg(fusion);
-
-    const auto device_prop = at::cuda::getCurrentDeviceProperties();
-
-    const int64_t sm_register_file_size =
-        static_cast<int64_t>(device_prop->regsPerBlock * sizeof(int));
-
-    auto persistent_buffer_info_entry =
-        HeuristicSummaryEntry<HeuristicCompileTime::PersistentBufferInfo>(
-            data_cache, [&fusion]() {
-              return std::make_unique<scheduler_utils::PersistentBufferInfo>(
-                  scheduler_utils::persistentBuffers(fusion));
-            });
-
-    const auto& persistent_buffer_info = persistent_buffer_info_entry.get();
-
-    auto persistent_buffer_size_info = scheduler_utils::persistentBufferSize(
-        fusion, runtime_info, persistent_buffer_info, data_cache);
-
-    // Note that projected buffer size can be zero
-    auto persistent_buffer_size =
-        persistent_buffer_size_info.projected_persistent_buffer_size == 0
-        ? persistent_buffer_size_info.persistent_buffer_size
-        : std::min(
-              persistent_buffer_size_info.persistent_buffer_size,
-              persistent_buffer_size_info.projected_persistent_buffer_size);
-
-    const int64_t device_multiprocessor_count =
-        (int64_t)device_prop->multiProcessorCount;
-
-    const auto available_persistent_buffer_size =
-        sm_register_file_size * device_multiprocessor_count;
-
-    if (persistent_buffer_size > available_persistent_buffer_size) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "not enough registers for persistence");
-      return false;
-    }
-
-    auto reduced_tv = ir_utils::getSoleProducerTv(reduction_tvs.at(0));
-
-    const int64_t vectorization_factor =
-        (int64_t)vectorize_helper::getVectorizationFactor(
-            runtime_info,
-            reduced_tv,
-            data_cache,
-            (int)reduced_tv->nDims() -
-                (int)properties.inner_most_dimension_ndims);
-
-    // Minimum required multi reduction factor.
-    const int64_t min_multi_reduction_factor = vectorization_factor *
-        normalization_scheduler_utils::PreferredLaunchConfig::kMinBdimx;
-
-    const int64_t required_sm_per_norm = ceilDiv(
-        persistent_buffer_size * min_multi_reduction_factor,
-        sm_register_file_size);
-
-    // If the persistence requires over half the device don't do grid
-    // persistence as we can't overlap the grid comms.
-    if (required_sm_per_norm >
-        scheduler_utils::safeDiv(device_multiprocessor_count, 2)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "requires over half GPU persistence.",
-          " required SMs per normalization: ",
-          required_sm_per_norm);
-      return false;
-    }
-
-    const bool is_cross_grid = required_sm_per_norm > 1;
-
-    std::optional<normalization_scheduler_utils::GridOuterNormalizationParams>
-        cross_grid_params;
-
-    if (is_cross_grid) {
-      // Don't try to be persistent unless at least 4-way vectorized
-      // as register usage is hard to control
-      // TODO: Is this necessary for block persistence as well?
-      if (vectorization_factor < 4) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent, "not enough vectorized");
-        return false;
-      }
-
-      // Make sure there's a valid grid persistence launch config
-      cross_grid_params =
-          normalization_scheduler_utils::getGridOuterNormalizationParams(
-              properties.total_reduction_numel,
-              properties.total_iteration_numel,
-              vectorization_factor,
-              persistent_buffer_size);
-
-      if (!cross_grid_params.has_value()) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::Persistent, "no valid launch config found");
-        return false;
-      }
-    }
-
-    NVF_ERROR(!is_cross_grid || cross_grid_params.has_value())
-
-    // Maximum number of iteration dimensions we can have and still be
-    // persistent.
-    const int64_t max_multi_reduction_factor = scheduler_utils::safeDiv(
-        is_cross_grid ? available_persistent_buffer_size
-                      : sm_register_file_size,
-        persistent_buffer_size);
-
-    // Don't go persistent if we can't fit the minimum multi reduction
-    // factor
-    if (max_multi_reduction_factor < min_multi_reduction_factor) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "Not enough threads.",
-          " Multi reduction factor, ",
-          max_multi_reduction_factor,
-          ", is smaller than minimum multi reduction factor, ",
-          min_multi_reduction_factor);
-      return false;
-    }
-
-    const int64_t max_used_sms = is_cross_grid
-        ? ceilDiv(
-              ceilDiv(properties.total_iteration_numel, vectorization_factor),
-              cross_grid_params->launch_params.bdimx()) *
-            cross_grid_params->launch_params.gdimy()
-        : ceilDiv(
-              properties.total_iteration_numel * persistent_buffer_size,
-              sm_register_file_size);
-
-    // Bandwidth suffers if the number of used SMs is small. This is
-    // particularly impactful in the case of cross grid, so at least
-    // half of the SMs are required to be used. In the case of cross
-    // block, keep using the existing heuristics for now.
-    if (is_cross_grid &&
-        max_used_sms <
-            scheduler_utils::safeDiv(device_multiprocessor_count, 2)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent,
-          "cross grid - not enough used SMs: ",
-          max_used_sms);
-      return false;
-    }
-
-    const int64_t device_max_threads_per_multiprocessor =
-        (int64_t)device_prop->maxThreadsPerMultiProcessor;
-    const int64_t min_fraction_of_sms =
-        scheduler_utils::safeDiv(device_multiprocessor_count, 8);
-    if (properties.total_reduction_numel >=
-            device_max_threads_per_multiprocessor * 4 && // Large reduction dim
-        max_used_sms < min_fraction_of_sms) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "not enough used SMs");
-      return false;
-    }
-
-    // The runtime kernel for grouped normal grid reductions is not
-    // well tuned, and it turned out to be quite difficult to get
-    // consistently better performances than non-persistent
-    // schedules. Disabled for now.
-    // TODO: Enable non-welford persistent reductions
-    if (is_cross_grid &&
-        std::any_of(
-            reduction_tvs.begin(),
-            reduction_tvs.end(),
-            [](TensorView* reduction_tv) {
-              return !reduction_tv->definition()->isA<WelfordOp>();
-            })) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "non-Welford not enabled yet");
-      return false;
-    }
-
-    // Had a hard time tuning on Titan RTX and V100 when the iteration
-    // space is not evenly divided by threads and thread blocks. It
-    // doesn't seem to be noticeably bad on A100, though. For now,
-    // disable the schedule if not evenly divisible on Titan RTX and
-    // V100, i.e., compute architecture version 7.
-    // TODO: Revisit
-    if (is_cross_grid &&
-        (properties.total_iteration_numel %
-             (vectorization_factor * cross_grid_params->launch_params.bdimx() *
-              cross_grid_params->launch_params.gdimx()) !=
-         0) &&
-        device_prop->major == 7) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::Persistent, "iteration not evenly divided");
-      return false;
-    }
-
-    return true;
   }
 };
 
@@ -2358,7 +750,7 @@ bool checkCanSchedule(
   //  it has to pass all the compile time checks to create a data cache for this
   //  fusion.
   if (!data_cache) {
-    if (!isConnectedFusionGraph(fusion)) {
+    if (!registry_utils::isConnectedFusionGraph(fusion)) {
       return false;
     }
     if (IterDomainGraph(fusion, /*allow_self_mapping=*/true).hasSelfMapping()) {
@@ -2458,31 +850,6 @@ std::optional<ScheduleHeuristic> SchedulerEntry::proposeHeuristics(
 
 size_t SchedulerEntryHash::operator()(const SchedulerEntry& se) const {
   return se.params()->hash();
-}
-
-std::string toString(ScheduleHeuristic sh) {
-  switch (sh) {
-    case ScheduleHeuristic::NoOp:
-      return "no-op";
-    case ScheduleHeuristic::PointWise:
-      return "pointwise";
-    case ScheduleHeuristic::Reduction:
-      return "reduction";
-    case ScheduleHeuristic::Persistent:
-      return "persistent";
-    case ScheduleHeuristic::Transpose:
-      return "transpose";
-    case ScheduleHeuristic::Matmul:
-      return "matmul";
-    default:
-      NVF_ERROR(false, "undefined schedule");
-  }
-  return "";
-}
-
-std::ostream& operator<<(std::ostream& os, ScheduleHeuristic sh) {
-  os << toString(sh);
-  return os;
 }
 
 namespace {
